feelpp-env/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feelpp-docker-factory"
version = "0.1.0"
description = "Docker image factory for Feel++ development environments"
authors = [
    {name = "Feel++ Team", email = "support@feelpp.org"}
]
readme = "DOCKER_FACTORY_EVALUATION.md"
requires-python = ">=3.9"
dependencies = [
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
feelpp-factory = "factory.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["factory*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
